# project-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Actor Platform API",
    "version": "0.0.1",
    "description": "API for managing actors, runs, schedules, storage, and webhooks on the actor automation platform."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "paths": {
...
    "/workspaces/{workspaceId}/datasets": {
      "get": {
        "summary": "List datasets",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paginated dataset list", "content": { "application/json": { "schema": { "type": "object", "properties": { "datasets": { "type": "array", "items": { "$ref": "#/components/schemas/Dataset" } }, "nextCursor": { "type": "string" } } } } } } }
      },
      "post": {
        "summary": "Create a dataset",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:write"] }],
        "parameters": [{ "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateStorageInput" } } } },
        "responses": { "201": { "description": "Dataset created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Dataset" } } } } }
      }
    },
    "/workspaces/{workspaceId}/datasets/{datasetId}": {
      "get": {
        "summary": "Get dataset by ID",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "datasetId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Dataset with item count", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DatasetWithCount" } } } } }
      },
      "delete": {
        "summary": "Delete a dataset",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:write"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "datasetId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Dataset deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/workspaces/{workspaceId}/datasets/{datasetId}/items": {
      "get": {
        "summary": "List dataset items",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "datasetId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paginated items", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/DatasetItem" } }, "nextCursor": { "type": "string" } } } } } } }
      },
      "post": {
        "summary": "Add item to dataset",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:write"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "datasetId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddItemInput" } } } },
        "responses": { "201": { "description": "Item added", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DatasetItem" } } } } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "SHA-256 hashed API token"
      }
    },
    "schemas": {
...
      "SuccessResponse": { "type": "object", "properties": { "success": { "type": "boolean" } }, "required": ["success"] },
...
      "Dataset": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "workspaceId": { "type": "string" }, "actorRunId": { "type": "string" },
          "ownerId": { "type": "string" }, "name": { "type": "string" }, "slug": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "DatasetWithCount": {
        "allOf": [
          { "$ref": "#/components/schemas/Dataset" },
          { "type": "object", "properties": { "_count": { "type": "object", "properties": { "items": { "type": "integer" } } } } }
        ]
      },
      "DatasetItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "datasetId": { "type": "string" }, "sequence": { "type": "integer" },
          "payload": {}, "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateStorageInput": {
        "type": "object",
        "required": ["name", "slug"],
        "properties": { "name": { "type": "string" }, "slug": { "type": "string" } }
      },
      "AddItemInput": {
        "type": "object",
        "required": ["payload"],
        "properties": {
          "payload": { "type": "object" },
          "sequence": { "type": "integer" }
        }
      },
...
    }
  }
}