# src/datasets.ts
import { Hono, Context } from "hono";
import { z } from "zod";
import { getPrisma } from "./config.js";
import type { Prisma } from "@prisma/client";
import { requireWorkspaceRole } from "./workspace-auth.js";
import { requireTokenScope } from "./token-scope.js";
import { assertStorageManageAccess } from "./storage-auth.js";

function getWorkspaceId(c: Context): string | Response {
  const id = c.req.param("workspaceId");
  if (!id) return c.json({ error: "Missing workspaceId in URL" }, 400);
  return id;
}

function getDatasetId(c: Context): string | Response {
  const id = c.req.param("datasetId");
  if (!id) return c.json({ error: "Missing datasetId in URL" }, 400);
  return id;
}

function getUserId(c: Context): string {
  return c.get("userId" as never) as string;
}

const createDatasetSchema = z.object({
  name: z.string().min(1, "name is required"),
  slug: z.string().min(1, "slug is required"),
});

const listQuerySchema = z.object({
  limit: z.coerce.number().int().min(1).max(100).default(10),
  cursor: z.string().optional(),
});

const addItemSchema = z.object({
  payload: z.record(z.unknown()),
  sequence: z.number().int().optional(),
});

const datasets = new Hono();

datasets.get("/", requireWorkspaceRole(["OWNER", "ADMIN", "MEMBER"]), requireTokenScope("storage:read"), async (c) => {
  const maybeWsId = getWorkspaceId(c);
  if (maybeWsId instanceof Response) return maybeWsId;
  const workspaceId: string = maybeWsId;

  const rawQuery = {
    limit: c.req.query("limit"),
    cursor: c.req.query("cursor"),
  };

  const parsed = listQuerySchema.safeParse(rawQuery);
  if (!parsed.success) {
    return c.json({ error: "Invalid query parameters", details: parsed.error.flatten() }, 400);
  }

  const { limit, cursor } = parsed.data;
  const prisma = getPrisma();

  try {
    const items = await prisma.dataset.findMany({
      where: { workspaceId },
      take: limit + 1,
      ...(cursor ? { cursor: { id: cursor }, skip: 1 } : {}),
      orderBy: { createdAt: "desc" },
    });

    const hasMore = items.length > limit;
    const datasetsList = hasMore ? items.slice(0, limit) : items;
    let nextCursor: string | undefined;
    if (hasMore) {
      const last = datasetsList.at(-1);
      if (last) nextCursor = last.id;
    }

    return c.json({ datasets: datasetsList, ...(nextCursor ? { nextCursor } : {}) });
  } catch {
    return c.json({ error: "Internal server error" }, 500);
  }
});

datasets.get("/:datasetId", requireWorkspaceRole(["OWNER", "ADMIN", "MEMBER"]), requireTokenScope("storage:read"), async (c) => {
  const maybeDatasetId = getDatasetId(c);
  if (maybeDatasetId instanceof Response) return maybeDatasetId;
  const datasetId: string = maybeDatasetId;
  const prisma = getPrisma();

  try {
    const dataset = await prisma.dataset.findUnique({
      where: { id: datasetId },
      include: { _count: { select: { items: true } } },
    });

    if (!dataset) {
      return c.json({ error: "Dataset not found" }, 404);
    }

    return c.json(dataset);
  } catch {
    return c.json({ error: "Internal server error" }, 500);
  }
});

datasets.get("/:datasetId/export", requireWorkspaceRole(["OWNER", "ADMIN", "MEMBER"]), requireTokenScope("storage:read"), async (c) => {
  const maybeDatasetId = getDatasetId(c);
  if (maybeDatasetId instanceof Response) return maybeDatasetId;
  const datasetId: string = maybeDatasetId;
  const prisma = getPrisma();

  try {
    const dataset = await prisma.dataset.findUnique({
      where: { id: datasetId },
      select: { id: true, name: true },
    });

    if (!dataset) {
      return c.json({ error: "Dataset not found" }, 404);
    }

    const items = await prisma.datasetItem.findMany({
      where: { datasetId },
      orderBy: { sequence: "asc" },
      select: { sequence: true, payload: true, createdAt: true },
    });

    const filename = `${dataset.name.replace(/[^a-zA-Z0-9_-]/g, "_")}.json`;
    c.header("Content-Disposition", `attachment; filename="${filename}"`);
    return c.json({ items, dataset: { id: dataset.id, name: dataset.name } });
  } catch {
    return c.json({ error: "Internal server error" }, 500);
  }
});

datasets.post("/", requireWorkspaceRole(["OWNER", "ADMIN", "MEMBER"]), requireTokenScope("storage:write"), async (c) => {
  const maybeWsId = getWorkspaceId(c);
  if (maybeWsId instanceof Response) return maybeWsId;
  const workspaceId: string = maybeWsId;

  let raw: unknown;
  try {
    raw = await c.req.json();
  } catch {
    return c.json({ error: "Invalid JSON body" }, 400);
  }

  const parsed = createDatasetSchema.safeParse(raw);
  if (!parsed.success) {
    return c.json({ error: "Invalid input", details: parsed.error.flatten() }, 400);
  }

  const { name, slug } = parsed.data;
  const ownerId = getUserId(c);
  const prisma = getPrisma();

  try {
    const dataset = await prisma.dataset.create({
      data: { workspaceId, name, slug, ownerId },
    });

    return c.json(dataset, 201);
  } catch {
    return c.json({ error: "Internal server error" }, 500);
  }
});

datasets.delete("/:datasetId", requireWorkspaceRole(["OWNER", "ADMIN", "MEMBER"]), requireTokenScope("storage:write"), async (c) => {
  const maybeWsId = getWorkspaceId(c);
  if (maybeWsId instanceof Response) return maybeWsId;
  const workspaceId: string = maybeWsId;
  const maybeDatasetId = getDatasetId(c);
  if (maybeDatasetId instanceof Response) return maybeDatasetId;
  const datasetId: string = maybeDatasetId;
  const prisma = getPrisma();

  try {
    const existing = await prisma.dataset.findUnique({
      where: { id: datasetId },
      select: { id: true, workspaceId: true, ownerId: true },
    });

    if (!existing) {
      return c.json({ error: "Dataset not found" }, 404);
    }

    if (existing.workspaceId !== workspaceId) {
      return c.json({ error: "Dataset not found" }, 404);
    }

    const hasAccess = await assertStorageManageAccess(c, datasetId, workspaceId, "dataset");
    if (!hasAccess) {
      return c.json({ error: "Forbidden: insufficient permissions" }, 403);
    }

    await prisma.dataset.delete({ where: { id: datasetId } });
    return c.json({ success: true });
  } catch {
    return c.json({ error: "Internal server error" }, 500);
  }
});

datasets.get("/:datasetId/items", requireWorkspaceRole(["OWNER", "ADMIN", "MEMBER"]), requireTokenScope("storage:read"), async (c) => {
  const maybeDatasetId = getDatasetId(c);
  if (maybeDatasetId instanceof Response) return maybeDatasetId;
  const datasetId: string = maybeDatasetId;

  const rawQuery = {
    limit: c.req.query("limit"),
    cursor: c.req.query("cursor"),
  };

  const parsed = listQuerySchema.safeParse(rawQuery);
  if (!parsed.success) {
    return c.json({ error: "Invalid query parameters", details: parsed.error.flatten() }, 400);
  }

  const { limit, cursor } = parsed.data;
  const prisma = getPrisma();

  try {
    const dataset = await prisma.dataset.findUnique({
      where: { id: datasetId },
      select: { id: true },
    });

    if (!dataset) {
      return c.json({ error: "Dataset not found" }, 404);
    }

    const items = await prisma.datasetItem.findMany({
      where: { datasetId },
      take: limit + 1,
      ...(cursor ? { cursor: { id: cursor }, skip: 1 } : {}),
      orderBy: { sequence: "asc" },
    });

    const hasMore = items.length > limit;
    const itemList = hasMore ? items.slice(0, limit) : items;
    let nextCursor: string | undefined;
    if (hasMore) {
      const last = itemList.at(-1);
      if (last) nextCursor = last.id;
    }

    return c.json({ items: itemList, ...(nextCursor ? { nextCursor } : {}) });
  } catch {
    return c.json({ error: "Internal server error" }, 500);
  }
});

datasets.post("/:datasetId/items", requireWorkspaceRole(["OWNER", "ADMIN", "MEMBER"]), requireTokenScope("storage:write"), async (c) => {
  const maybeWsId = getWorkspaceId(c);
  if (maybeWsId instanceof Response) return maybeWsId;
  const workspaceId: string = maybeWsId;
  const maybeDatasetId = getDatasetId(c);
  if (maybeDatasetId instanceof Response) return maybeDatasetId;
  const datasetId: string = maybeDatasetId;

  let raw: unknown;
  try {
    raw = await c.req.json();
  } catch {
    return c.json({ error: "Invalid JSON body" }, 400);
  }

  const parsed = addItemSchema.safeParse(raw);
  if (!parsed.success) {
    return c.json({ error: "Invalid input", details: parsed.error.flatten() }, 400);
  }

  const { payload, sequence } = parsed.data;
  const prisma = getPrisma();

  try {
    const dataset = await prisma.dataset.findUnique({
      where: { id: datasetId },
      select: { id: true, workspaceId: true, ownerId: true },
    });

    if (!dataset) {
      return c.json({ error: "Dataset not found" }, 404);
    }

    if (dataset.workspaceId !== workspaceId) {
      return c.json({ error: "Dataset not found" }, 404);
    }

    const hasAccess = await assertStorageManageAccess(c, datasetId, workspaceId, "dataset");
    if (!hasAccess) {
      return c.json({ error: "Forbidden: insufficient permissions" }, 403);
    }

    const maxSeq = await prisma.datasetItem.findFirst({
      where: { datasetId },
      orderBy: { sequence: "desc" },
      select: { sequence: true },
    });

    const nextSeq = sequence ?? (maxSeq ? maxSeq.sequence + 1 : 0);

    const itemPayload: Prisma.InputJsonValue = payload as Prisma.InputJsonValue;
    const item = await prisma.datasetItem.create({
      data: { datasetId, sequence: nextSeq, payload: itemPayload },
    });

    return c.json(item, 201);
  } catch {
    return c.json({ error: "Internal server error" }, 500);
  }
});

export { datasets };